requirements.txt
psycopg2-binary
tabulate
pyinstaller